CI pipeline (#88)
* ci: initial commit of the CI pipeline

Install semantic-release, husky, commitlint
Create the Github Actions workflow

* chore: fix docker-compose.yml for WSL

WSL filesystem seems not to be compatible with Mongo, so we use proper Docker volumes instead

* refactor: fix lint errors

* style: npm run format

* style: use prettier-eslint to make lint & format compatible with eachother

* chore: check lint & format during pre-commit hook

* chore: set a maximum of 55 eslint warnings

Currently there are 55 warnings. We enforce this limit to prevent from adding new warnings.
After a refactoring of these existing warnings, the limit should be lowered to 0.

* feat: reset version to 1.0.0

Co-authored-by: Thomas Vaillant <thomas.vaillant-ext@sanofi.com>